RUM-6698 CR feedback
